Move RootConfig, RootPatchSettings to a model module
